Add version 3 of package with old version 2 contents
Due to the way golang handles major versions > v1, the package path *must* contain the major version in the import/module path. This means that importing module version 2 did not work because there was no v2 in the path.

To get around this while still maintaining semantic versioning, a package version 3 was created in top-level directory 'v3'. This means import paths for v3 tags will now contain the 'v3' suffix as expected.

Additional enhancements were made to the package for v3. Since a breaking change was going to be introduced anyway, some of the package files were reorganized to remove the verbose 'pkg/eolib/...' imports - the directories are all at the top of v3 together now.